Reorganized the 'tests' directory into smaller files stored in multiple directories. Minor changes were added to the tests
